requirements.txt
python-telegram-bot==21.4
starlette==0.37.2
uvicorn==0.29.0
httpx==0.27.0
nest_asyncio==1.6.0
